ready for demo

+++ server/rest/draft_project/draft_projects_service.py
@@ -1,45 +1,50 @@
 from db.models import ProjectDraft,Project
-from errors import NotFound
 from mongoengine.queryset.visitor import Q
 from ..utils import utils
+from werkzeug.exceptions import NotFound
 
-model = ProjectDraft
 
-def get_draft_project(id):
-    return utils.return_document_by_id(model, dict(project_id=id), fields_to_exclude=['id','created','valid'])
+def get_draft_project(project_id):
+    draft_projects=utils.get_documents_by_query(ProjectDraft, dict(project_id=project_id),('id','created','valid'))
+    if draft_projects.first():
+        return draft_projects.as_pymongo()[0]
+    raise NotFound(description=f"Draft Project: {project_id} not found!")
 
 def create_draft_project(data):
     for req_field in ['project_id', 'name','version']:
         if not data.get(req_field):
             message = f"{req_field} is mandatory in ProjectDraft objects"
-            status = 500
-            return dict(message=message),status
+            status = 400
+            return [message],status
     if Project.objects(project_id=data['project_id']).first():
         message= f"An existing project with id: {data['project_id']} already exists"
-        status = 500
+        status = 409
     elif ProjectDraft.objects(project_id=data['project_id']).first():
         message = f"An existing draft project with id: {data['project_id']} already exists"
-        status = 500
+        status = 409
     else:
         new_draft_project = ProjectDraft(**data)
         new_draft_project.save()
         message = f"Draft project with id: {new_draft_project.project_id} correctly created"
         status = 201
-    return dict(message=message),status
+    return [message],status
 
-def update_draft_project(id, data):
-    project_to_update = model.objects(project_id=id)
-    project_to_update.update(**data)
-    message = f"Draft project {id} correctly updated"
-    return dict(message=message), 201
+def update_draft_project(project_id, data):
+    projects_to_update = utils.get_documents_by_query(ProjectDraft, dict(project_id=project_id))
+    if not project_to_update.first():
+        raise NotFound(description=f"Draft Projects: {project_id} not found!")
+    project_to_update = ProjectDraft.objects(project_id=project_id)
+    projects_to_update.update(**data)
+    message = f"Draft project {project_id} correctly updated"
+    return [message], 201
 
 
 def get_draft_projects(offset=0,limit=20,
                 filter=None,sort_order=None):
     if filter:
-        projects= model.objects((Q(name__icontains=filter) | Q(name__iexact=filter))).exclude('id','created')
+        projects= ProjectDraft.objects((Q(name__icontains=filter) | Q(name__iexact=filter))).exclude('id','created')
     else:
-        projects = model.objects().exclude('id','created')
+        projects = ProjectDraft.objects().exclude('id','created')
     if sort_order:
         sort_column = "name"
         sort = '-'+sort_column if sort_order == 'desc' else sort_column
@@ -47,8 +52,10 @@ def get_draft_projects(offset=0,limit=20,
     return projects.count(), projects[int(offset):int(offset)+int(limit)]
 
 
-def delete_draft_project(id):
-    project_to_delete = get_draft_project(id)
+def delete_draft_project(project_id):
+    project_to_delete=utils.get_documents_by_query(ProjectDraft,dict(project_id=project_id)).first()
+    if not project_to_delete:
+        raise NotFound(description=f"Draft Projects: {project_id} not found!")
     project_to_delete.delete()
-    message= f"Draft Project with id: {id} successfully deleted"
-    return dict(message=message), 201
+    message= f"Draft Project with id: {project_id} successfully deleted"
+    return [message], 201
